pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aria-agent"
version = "0.1.0"
description = "Autonomous AI assistant for macOS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.0.0",
    "rumps>=0.4.0",
    "pyautogui>=0.9.54",
    "pillow>=10.0.0",
    "pyobjc-core>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "numpy>=1.24.0",
    "sounddevice>=0.4.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
wake-word = ["pvporcupine>=3.0.0", "pyaudio>=0.2.14"]

[project.scripts]
aria = "aria.main:main"

[tool.setuptools.packages.find]
include = ["aria*"]
